Add a slice transform
- for strings
- for arrays
